Move initialization data model
