+++ src/xcode/ENA/ENA/Source/Scenes/QRScanner/Model/QRCodeResult.swift
@@ -17,17 +17,6 @@ struct CertificateResult {
 	let certificate: HealthCertificate
 }
 
-struct TicketValidationInitializationData: Codable {
-	let `protocol`: String
-	let protocolVersion: String
-	let serviceIdentity: String
-	let privacyUrl: String
-	let token: String
-	let consent: String
-	let subject: String
-	let serviceProvider: String
-}
-
 enum QRCodeParserError: Error, Equatable {
 	case scanningError(QRScannerError)
 	case checkinQrError(CheckinQRScannerError)
