+++ src/xcode/ENA/ENA/Source/Services/TicketValidation/Models/TicketValidationResult+Fake.swift
