+++ src/xcode/ENA/ENA/Source/Services/TicketValidation/Models/ValidationConditions.swift
